implemented ogl camera + first rotation of scene
scene uses internally the OpenGL camera, set via gluLookAt(...) to
change the view.
that way, not the scene itself is rotated, but the camera around the
scene.
the camera is controlled via the KBX_Camera class, which supports
rotation around a given target point (where the camera looks at) and
zoom in/out to/from the target point (not yet implemented).

the scene object will call the camera's updateView() method to set the
camera to the specified position.
this is done in the scene's display() method.