corrected rotation code of KBX_Object
now, you can perform several rotations successively.
normally, the coordinate system of OpenGL would rotate with the object,
thus giving the programmer a headache when trying to align an object in
the right way.
here, the rotations are pushed to a queue while being transformed to the
local frame of reference.
that way, the programmer can do something like
    # rotate clockwise
    # rotate forward
and the orientation will always be in the direction of the first
rotation, i.e. in a fixed direction (like you would expect, if you are a
fixed spectator).