changed mouse behaviour a bit: right btn to rotate
now the right button is used to rotate the board, while the left button
is still used to select and the middle button is used to zoom.

additionally there is no difference in using the lower or the upper half
of the screen any more, as this led to a bit unexpected behaviour.

user thought: "why does it now move in the other direction, when the
direction of the mouse is still the same?"