started implementing controller
a controller connects a model (in this case the engine) with a view
(here the gui).
the controller will hold the information, which die will be connected to
which id in the abstract representation of the game (engine).
additionally, the event handlers will be modified, so that they emit
special SDL events directed to the controller, which specify, for
example, that a die has been selected.
that way, the controller can control the gameflow (let the computer
move, let the human move, etc) but doesnt have to listen to every raw
event coming from the gui, but listens only to events that have been
filtered and preprocessed by the event handlers.