more on texture loading, but STILL NOT WORKING
however, the web suggests that QGLWidget::bindTexture
does not work yet (known issue).
because of that, we will implement it the old fashioned
way by calling the according opengl functions directly.
bitmap data will be loaded via QPixmap (tested, that works).
texture data, however, seems to get lost on the way to
the graphics card...